Safari-609.1.11.1/LayoutTests/http/tests/cache/disk-cache/disk-cache-remove-several-pending-writes-expected.txt

Make sure that we properly remove cached entry if the entry is no longer cacheable after revalidation

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


200: {"version":1}
200: {"version":1}
404: {"not":"found"}
200: {"version":2}
200: {"version":2}
PASS bugReproduced is false
PASS successfullyParsed is true

TEST COMPLETE
